moved Location operations to the new DAO format.
